pyproject.toml
[project]
name = "news-feeder"
version = "0.1.0"
description = "RSSフィードを取得して日本語に翻訳し、Gmail でメール配信する Python スクリプト"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0",
    "requests>=2.28",
    "pyyaml>=6.0",
]

[project.scripts]
news-feeder = "news_feeder.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/news_feeder"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
]
